Update sample-press-release.txt
Paraphrasing the second paragraph to the following :

A landing page differs from a website's homepage or a product page in that it is purposefully created with a specific objective in mind. It hones in on a singular offer or message, eliminating any potential distractions that could steer visitors away from that primary goal. As a result, landing pages play a crucial role in digital marketing campaigns, greatly enhancing the likelihood of converting visitors into valuable customers.